Refactor et renommage + intégration setup
